+++ backend/src/data-source.ts
@@ -11,8 +11,8 @@ export const AppDataSource = new DataSource({
     type: "postgres",
     host: "localhost",
     port: 5432,
-    username: "test",
-    password: "test",
+    username: "postgres",
+    password: "arthur",
     database: "test",
     synchronize: true,
     logging: true,
